anything up to Segments
