+++ src/arg_desc.jl
@@ -8,20 +8,20 @@ const _arg_desc = KW(
     :label              => (AStr, "The label for a series, which appears in a legend. If empty, no legend entry is added."),
     :seriescolor        => (ColorType, "The base color for this series. `:auto` (the default) will select a color from the subplot's `color_palette`, based on the order it was added to the subplot. Also describes the colormap for surfaces."),
     :seriesalpha        => (Real, "The alpha/opacity override for the series. `nothing` (the default) means it will take the alpha value of the color."),
-    :seriestype         => (Symbol, "This is the identifier of the type of visualization for this series. Choose from $(_allTypes) or any series recipes which are defined."),
-    :linestyle          => (Symbol, "Style of the line (for path and bar stroke). Choose from $(_allStyles)"),
+    :seriestype         => (Symbol, "This is the identifier of the type of visualization for this series. Choose from $(Commons._allTypes) or any series recipes which are defined."),
+    :linestyle          => (Symbol, "Style of the line (for path and bar stroke). Choose from $(Commons._allStyles)"),
     :linewidth          => (Real, "Width of the line (in pixels)."),
     :linecolor          => (ColorType, "Color of the line (for path and bar stroke). `:match` will take the value from `:seriescolor`, (though histogram/bar types use `:black` as a default)."),
     :linealpha          => (Real, "The alpha/opacity override for the line. `nothing` (the default) means it will take the alpha value of linecolor."),
     :fillrange          => (Union{Real,AVec}, "Fills area between fillrange and `y` for line-types, sets the base for `bar`, `sticks` types, and similar for other types."),
     :fillcolor          => (ColorType, "Color of the filled area of path or bar types. `:match` will take the value from `:seriescolor`."),
     :fillalpha          => (Real, "The alpha/opacity override for the fill area. `nothing` (the default) means it will take the alpha value of fillcolor."),
-    :markershape        => (Union{Symbol,Shape,AVec}, "Choose from $(_allMarkers)."),
+    :markershape        => (Union{Symbol,Shape,AVec}, "Choose from $(Commons._allMarkers)."),
     :fillstyle          => (Symbol, "Style of the fill area. `nothing` (the default) means solid fill. Choose from :/, :\\, :|, :-, :+, :x."),
     :markercolor        => (ColorType, "Color of the interior of the marker or shape. `:match` will take the value from `:seriescolor`."),
     :markeralpha        => (Real, "The alpha/opacity override for the marker interior. `nothing` (the default) means it will take the alpha value of markercolor."),
     :markersize         => (Union{Real,AVec}, "Size (radius pixels) of the markers."),
-    :markerstrokestyle  => (Symbol, "Style of the marker stroke (border). Choose from $(_allStyles)."),
+    :markerstrokestyle  => (Symbol, "Style of the marker stroke (border). Choose from $(Commons._allStyles)."),
     :markerstrokewidth  => (Real, "Width of the marker stroke (border) in pixels."),
     :markerstrokecolor  => (ColorType, "Color of the marker stroke (border). `:match` will take the value from `:foreground_color_subplot`."),
     :markerstrokealpha  => (Real, "The alpha/opacity override for the marker stroke (border). `nothing` (the default) means it will take the alpha value of markerstrokecolor."),
@@ -125,7 +125,7 @@ const _arg_desc = KW(
     :colorbar_tickfontfamily     => (Union{AStr,Symbol}, "String or Symbol. Font family of colorbar tick labels."),
     :colorbar_tickfontsize       => (Integer, "Font pointsize of colorbar tick entries."),
     :colorbar_tickfontcolor      => (ColorType, "Font color of colorbar tick entries."),
-    :colorbar_scale              => (Symbol, "Scale of the colorbar axis. Choose from $(_allScales)."),
+    :colorbar_scale              => (Symbol, "Scale of the colorbar axis. Choose from $(Commons._allScales)."),
     :colorbar_formatter          => (Union{Function,Symbol}, "Choose from (:scientific, :plain, :none, :auto), or a method which converts a number to a string for tick labeling."),
     :legend_font                 => (Font, "Font of legend items."),
     :legend_titlefont            => (Font, "Font of the legend title."),
@@ -146,7 +146,7 @@ const _arg_desc = KW(
     :bottom_margin               => (Union{Tuple,Real,Symbol}, "Specifies the extra padding on the bottom of the subplot (`:match` matches `:margin`)."),
     :subplot_index               => (Integer, "Internal (not set by user). Specifies the index of this subplot in the Plot's `plt.subplot` list."),
     :colorbar_title              => (AStr, "Title of colorbar."),
-    :framestyle                  => (Symbol, "Style of the axes frame. Choose from $(_allFramestyles)."),
+    :framestyle                  => (Symbol, "Style of the axes frame. Choose from $(Commons._allFramestyles)."),
     :camera                      => (NTuple{2,Real}, "Sets the view angle (azimuthal, elevation) for 3D plots."),
 
     # axis args
@@ -158,7 +158,7 @@ const _arg_desc = KW(
                                     `:symmetric` sets the limits to be symmetric around zero.
                                     Set `widen=true` to widen the specified limits (as occurs when lims are not specified)."""),
     :ticks                       => (TicksType, "Tick values, (tickvalues, ticklabels), `:auto`/`true`, `:none`/`false`/`nothing` (ticks disabled), or `:native` (tells backend to calculate ticks by itself; good idea for interactive backends with mouse zooming)."),
-    :scale                       => (Symbol, "Scale of the axis. Choose from $(_allScales)."),
+    :scale                       => (Symbol, "Scale of the axis. Choose from $(Commons._allScales)."),
     :rotation                    => (Real, "Degrees rotation of tick labels."),
     :flip                        => (Bool, "Should we flip (reverse) the axis ?"),
     :formatter                   => (Union{Symbol,Function}, "Choose from (:scientific, :plain or :auto), or a method which converts a number to a string for tick labeling."),
@@ -182,13 +182,13 @@ const _arg_desc = KW(
     :grid                        => (Union{Bool,Symbol,AStr}, "Show the grid lines ? `true`, `false`, `:show`, `:hide`, `:yes`, `:no`, `:x`, `:y`, `:z`, `:xy`, ..., `:all`, `:none`, `:off`."),
     :foreground_color_grid       => (ColorType, "Color of grid lines (`:match` matches `:foreground_color_subplot`)."),
     :gridalpha                   => (Real, "The alpha/opacity override for the grid lines."),
-    :gridstyle                   => (Symbol, "Style of the grid lines. Choose from $(_allStyles)."),
+    :gridstyle                   => (Symbol, "Style of the grid lines. Choose from $(Commons._allStyles)."),
     :gridlinewidth               => (Real, "Width of the grid lines (in pixels)."),
     :foreground_color_minor_grid => (ColorType, "Color of minor grid lines (`:match` matches `:foreground_color_subplot`)."),
     :minorgrid                   => (Bool, "Adds minor grid lines and ticks to the plot. Set minorticks to change number of gridlines."),
     :minorticks                  => (Integer, "Number of minor intervals between major ticks."),
     :minorgridalpha              => (Real, "The alpha/opacity override for the minorgrid lines."),
-    :minorgridstyle              => (Symbol, "Style of the minor grid lines. Choose from $(_allStyles)."),
+    :minorgridstyle              => (Symbol, "Style of the minor grid lines. Choose from $(Commons._allStyles)."),
     :minorgridlinewidth          => (Real, "Width of the minor grid lines (in pixels)."),
     :tick_direction              => (Symbol, "Direction of the ticks. Choose from (`:in`, `:out`, `:none`)."),
     :showaxis                    => (Union{Bool,Symbol,AStr}, "Show the axis. `true`, `false`, `:show`, `:hide`, `:yes`, `:no`, `:x`, `:y`, `:z`, `:xy`, ..., `:all`, `:off`."),
