+++ src/axes_utils.jl
@@ -522,4 +522,3 @@ function axis_drawing_info_3d(sp, letter)
     )
 end
 
-reverse_if(x, cond) = cond ? reverse(x) : x
