+++ src/plotattr.jl
@@ -10,7 +10,7 @@ attrtypes() = join(keys(_attribute_defaults), ", ")
 attributes(attrtype::Symbol) = sort(collect(keys(_attribute_defaults[attrtype])))
 
 function lookup_aliases(attrtype::Symbol, attribute::Symbol)
-    attribute = get(_keyAliases, attribute, attribute)
+    attribute = get(Commons._keyAliases, attribute, attribute)
     attribute ∈ keys(_attribute_defaults[attrtype]) && return attribute
     error("There is no attribute named $attribute in $attrtype")
 end
@@ -30,22 +30,22 @@ function plotattr()
         @warn "Fuzzy finding of attributes is disabled in notebooks."
         return
     end
-    attr = Symbol(JLFzf.inter_fzf(collect(Plots._all_args), "--read0", "--height=80%"))
+    attr = Symbol(JLFzf.inter_fzf(collect(Commons._all_args), "--read0", "--height=80%"))
     letter = ""
-    attrtype = if attr ∈ _all_series_args
+    attrtype = if attr ∈ Commons._all_series_args
         "Series"
-    elseif attr ∈ _all_subplot_args
+    elseif attr ∈ Commons._all_subplot_args
         "Subplot"
-    elseif attr ∈ _lettered_all_axis_args
-        if attr ∉ _all_axis_args
+    elseif attr ∈ Commons._lettered_all_axis_args
+        if attr ∉ Commons._all_axis_args
             letters = collect(String(attr))
             letter = first(letters)
             attr = Symbol(join(letters[2:end]))
         end
         "Axis"
-    elseif attr ∈ _all_plot_args
+    elseif attr ∈ Commons._all_plot_args
         "Plot"
-    elseif attr ∈ _all_magic_args
+    elseif attr ∈ Commons._all_magic_args
         "Magic"
     else
         "Unknown"
@@ -69,7 +69,7 @@ end
 
 function plotattr(attribute::AbstractString)
     attribute = Symbol(attribute)
-    attribute = get(_keyAliases, attribute, attribute)
+    attribute = get(Commons._keyAliases, attribute, attribute)
     for (k, v) in _attribute_defaults
         attribute ∈ keys(v) && return plotattr(k, attribute)
     end
@@ -83,7 +83,7 @@ function plotattr(attrtype::Symbol, attribute::Symbol)
     attribute = lookup_aliases(attrtype, attribute)
     type, desc = _arg_desc[attribute]
     def = _attribute_defaults[attrtype][attribute]
-    aliases = if (al = Plots.aliases(attribute)) |> length > 0
+    aliases = if (al = Plots.Commons.aliases(attribute)) |> length > 0
         "Aliases: " * string(Tuple(al)) * ".\n\n"
     else
         ""
